refactor: extract ModalState to a context file
