+++ src/contexts/ModalContext.tsx
@@ -1,0 +1,6 @@
+export enum ModalState {
+  HIDDEN = 'HIDDEN',
+  LOGIN = 'LOGIN',
+  RESET = 'RESET',
+  SIGNUP = 'SIGNUP'
+}
